parking-control/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.0
opencv-python==4.9.0.80
pillow==10.2.0 